+++ src/utils/Average.c
@@ -1,7 +1,9 @@
 #include "../../include/Average.h"
 
-int Average(int list[], int size) {
-  int sum = 0;
+
+// this finds the sum of a list of numbers and divides it by the size to find the average 
+double Average(int list[], int size) {
+  double sum = 0;
   for (int i =0; i < size; i++) {
     sum += list[i];
   }
